Modified: ft_min() and ft_max(): make type double

+++ src/utils.c
@@ -6,7 +6,7 @@
 /*   By: squinn <squinn@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/04 17:06:13 by squinn            #+#    #+#             */
-/*   Updated: 2025/08/07 18:36:02 by squinn           ###   ########.fr       */
+/*   Updated: 2025/08/08 11:58:38 by squinn           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,13 +24,13 @@ t_point new_point(double x, double y, double z, uint32_t color) {
   return point;
 }
 
-int ft_min(int a, int b) {
+double ft_min(double a, double b) {
   if (a < b)
     return (a);
   return (b);
 }
 
-int ft_max(int a, int b) {
+double ft_max(double a, double b) {
   if (a > b)
     return (a);
   return (b);
